rename: fsd 구조 수정
