+++ app/login/pages.tsx
@@ -1,0 +1,5 @@
+import { LoginPage } from '@/views/login'
+
+export default function Page() {
+  return <LoginPage />
+}
