+++ src/entities/calendar/index.ts
@@ -1,1 +1,9 @@
 export { useScheduleStore } from './model/scheduleStore'
+export {
+  deleteSchedule,
+  getCalendarList,
+  getScheduleDetail,
+  postSchedule,
+  putSchedule,
+} from './api/calendarApi'
+export type { Schedule, ScheduleResponse } from './api/types'
